Change script to name graphs with folder name
Added the live data, and output graphs.

Debugged the script on live data from Blackrock Forest Trip Sat 10/15/16 to 10/16/16. The biggest change is giving the graph the same name as the folder in which the data is contained. For this reason, it is vital that the folder containing the data is given the unique device name.